fix placement of buttons on main screen
