+++ Front_End/app/onboarding/quiz/index.tsx
@@ -1,23 +1,40 @@
-import React, { useState } from 'react';
-import { View, StyleSheet } from 'react-native';
 import PrimaryButton from '@/components/PrimaryButton';
-import StepOne from './screens/StepOne';
 import QuizProgressBar from '@/components/ProgressBar';
-import StepTwo from './screens/StepTwo';
+import { api } from '@/services/api';
 import { useRouter } from 'expo-router';
+import React, { useState } from 'react';
+import { Alert, StyleSheet, View } from 'react-native';
+import StepOne from './screens/StepOne';
+import StepTwo from './screens/StepTwo';
 
 export default function QuizScreen() {
   const router = useRouter();
   const [step, setStep] = useState(1);
   const [stepTwoCompleted, setStepTwoCompleted] = useState(false);
   const [selectionId, setSelectionId] = useState<string | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
 
-  const handleContinue = () => {
+  const handleContinue = async () => {
     if (step === 1) {
       setStep(2);
       setStepTwoCompleted(false);
     } else if (step === 2 && stepTwoCompleted) {
-      router.push('/plaid-connection');
+      try {
+        setIsLoading(true);
+        // Submit quiz data
+        await api.submitQuiz({
+          user_id: 'temp-user-id', // Replace with actual user ID from auth
+          answers: {
+            stepOne: selectionId,
+            stepTwo: stepTwoCompleted
+          }
+        });
+        router.push('/plaid-connection');
+      } catch (error) {
+        Alert.alert('Error', 'Failed to submit quiz. Please try again.');
+      } finally {
+        setIsLoading(false);
+      }
     }
   };
 
@@ -44,6 +61,8 @@ export default function QuizScreen() {
         title="Continue"
         onPress={handleContinue}
         disabled={step === 1 ? !selectionId : !stepTwoCompleted}
+        // TO DO: ADD LOADING WHEN NEEDED
+        // loading={isLoading}
       />
     </View>
   );
